requirements.txt
discord.py==2.7.1
python-dotenv==1.2.2
aiohttp==3.13.5
aiosqlite==0.20.0
idna==3.11
Pillow==11.3.0
moviepy==2.2.1
imageio==2.37.3
imageio-ffmpeg==0.6.0
aiosqlite==0.20.0
